refactor: simplify save error notifications per review
Use Refine default error toasts with duration: 0 in app.tsx; drop custom errorNotification helpers.
