+++ packages/admin/vite.config.ts
@@ -69,7 +69,6 @@ export default defineConfig({
             shared: path.resolve(__dirname, './src/shared'),
             acl: path.resolve(__dirname, './src/acl.ts'),
             dataProvider: path.resolve(__dirname, './src/dataProvider.ts'),
-            notificationProvider: path.resolve(__dirname, './src/notificationProvider.ts'),
             assets: path.resolve(__dirname, './src/assets'),
             '@feed/shared/planning': path.resolve(__dirname, '../shared/src/planning/index.ts'),
             '@feed/shared': path.resolve(__dirname, '../shared/src/index.ts')
